DataTypeInserts, VariableFieldInserts, ReportInserts, puse required y ? en entities

+++ Sistema-de-rendicion-de-gastos/Application/DTO/Response/RealReport.cs
@@ -21,7 +21,7 @@ namespace Application.DTO.Response
 
         public string Description { get { return _report.Description; } set { _report.Description = value; } }
 
-        public int Amount { get { return _report.Amount; } set { _report.Amount = value; } }
+        public double Amount { get { return _report.Amount; } set { _report.Amount = value; } }
 
         //public ICollection<ReportTracking> Trackings { get { return _report.Trackings; } set { _report.Trackings = value; } }
         //public ICollection<VariableField> Fields { get { return _report.ReportId; } set { _report.Fields = value; } }
